The instructor demonstrates every exercise
All five exercises are now demonstrated by the AI pilot through a Controller,
and every aeroplane taught an exercise flies it rather than a sample: turns
(16), the take-off (5), the approach (6), the climb and descent (5) and the
stall (6). Each flies every stage to an empty debrief, hands over, and takes
the controls back - 0.0083 over and 0.0000 back, where a hand moves 0.0207.

What kept the item open was that two exercises were not flown by a controller
at all. The climb and the stall drove an autopilot directly and reached into
the controls it returned; the stall closed the throttle by hand on the way in
and opened it wide on the way out. There is nothing to hand over in that.

The instructor asks for a speed now, not for a throttle. Asking for a speed
below the stall closes the throttle and holds the height by raising the nose;
asking for one above the climbing speed opens it and puts the nose down. The
speed asked for has to clear the one the lesson waits on - half as much again
as the stall speed is 72 knots in a Cessna and the stage ends at 76, so she
settled underneath it and the stage never ended.

Also makes a float-to-double promotion explicit in the CLI, which clang
rejects and GCC does not.

+++ src/frontend/cli/main.cpp
@@ -577,7 +577,7 @@ int stay(glideslope::platform::UdpSocket& socket,
                 if (a.index == state->your_aircraft) {
                     if (std::abs(static_cast<double>(a.roll_deg)) >
                         std::abs(roll_seen_deg)) {
-                        roll_seen_deg = a.roll_deg;
+                        roll_seen_deg = static_cast<double>(a.roll_deg);
                     }
                 }
             }
